requirements/requirements.txt
attrs==17.4.0
boardgamegeek==0.13.2
flask==0.12.2
flask-injector==0.10.1
flask-sqlalchemy==2.3.2
injector==0.13.4
marshmallow==2.15.0
sqlalchemy==1.2.7
